service_auth/requirements.txt
fastapi==0.115.12
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
uvicorn[standard]==0.34.0

